requirements.txt
streamlit
phidata 
python-dotenv
duckduckgo-search
google-genai
google-generativeai